feat: refactor gateway connector architecture to use network format
- Change connector_name from DEX format (orca/clmm) to network format (solana-mainnet-beta)
- Add dex_name field for DEX protocol name (e.g., "orca", "meteora")
- Rename dex parameter to dex_name throughout for clarity
- Combined GatewaySwap and GatewayLp into unified Gateway class
- Update gateway swap command to use network format and auto-fetch default swap provider
- Update gateway LP and pool commands for new architecture
- Update all tests for new parameter naming

Architecture:
- connector_name: Network identifier (e.g., "solana-mainnet-beta")
- dex_name: DEX protocol name (e.g., "orca", "jupiter")
- trading_type: Pool type (e.g., "clmm", "amm", "router")
