+++ hummingbot/client/ui/parser.py
@@ -155,11 +155,11 @@ def load_parser(hummingbot: "HummingbotApplication", command_tabs) -> ThrowingAr
         "swap",
         help="Swap tokens")
     gateway_swap_parser.add_argument("connector", nargs="?", default=None,
-                                     help="Connector name/type (e.g., jupiter/router)")
+                                     help="Network (e.g., solana-mainnet-beta, ethereum-mainnet)")
     gateway_swap_parser.add_argument("args", nargs="*",
                                      help="Arguments: [base-quote] [side] [amount]. "
                                           "Interactive mode if not all provided. "
-                                          "Example: gateway swap uniswap ETH-USDC BUY 0.1")
+                                          "Example: gateway swap solana-mainnet-beta SOL-USDC BUY 0.1")
     gateway_swap_parser.set_defaults(func=hummingbot.gateway_swap)
 
     gateway_token_parser = gateway_subparsers.add_parser("token", help="View or update token information")
