+++ test/hummingbot/strategy_v2/executors/lp_executor/test_data_types.py
@@ -37,14 +37,16 @@ class TestLPExecutorConfig(TestCase):
         config = LPExecutorConfig(
             id="test-1",
             timestamp=1234567890,
-            connector_name="meteora/clmm",
+            connector_name="solana-mainnet-beta",
+            dex_name="meteora",
             trading_pair="SOL-USDC",
             pool_address="pool123",
             lower_price=Decimal("100"),
             upper_price=Decimal("110"),
         )
         self.assertEqual(config.type, "lp_executor")
-        self.assertEqual(config.connector_name, "meteora/clmm")
+        self.assertEqual(config.connector_name, "solana-mainnet-beta")
+        self.assertEqual(config.dex_name, "meteora")
         self.assertEqual(config.trading_pair, "SOL-USDC")
         self.assertEqual(config.pool_address, "pool123")
         self.assertEqual(config.lower_price, Decimal("100"))
@@ -62,7 +64,8 @@ class TestLPExecutorConfig(TestCase):
         config = LPExecutorConfig(
             id="test-2",
             timestamp=1234567890,
-            connector_name="meteora/clmm",
+            connector_name="solana-mainnet-beta",
+            dex_name="meteora",
             trading_pair="SOL-USDC",
             pool_address="pool456",
             lower_price=Decimal("90"),
@@ -89,7 +92,8 @@ class TestLPExecutorConfig(TestCase):
             config = LPExecutorConfig(
                 id=f"test-side-{side}",
                 timestamp=1234567890,
-                connector_name="meteora/clmm",
+                connector_name="solana-mainnet-beta",
+                dex_name="meteora",
                 trading_pair="SOL-USDC",
                 pool_address="pool",
                 lower_price=Decimal("100"),
